Initial commit
This project is incomplete with processor.h, but it is the part of the team project I completed, with a loader.h to load text from a file into a queue and a classes.h file containing blueprints for our objects and the main event loop.  Processor.h loads events into the priority queue so the event loop can process the events.